feat(campaigns): Denormalize `allCampaigns` for efficient querying
Adds a denormalized `allCampaigns` field to the Promoter and Campaign types. This array will store the campaign ID the promoter is associated with, along with any other campaigns they are linked to.

This change enables more efficient querying of promoters across multiple campaigns, particularly in scenarios where a promoter might be associated with more than one campaign.

Additionally, `associatedAdmins` has been added to the Campaign type to allow for admin association with campaigns.