feat: Add skip proof requirement option to posts
Introduces a new `skipProofRequirement` boolean field to the `Post` and `PostAssignment` interfaces. This allows organizers to create posts where proof of completion is not mandatory for promoters.

Updates the `CreatePost` and `EditPostModal` components to include this new option in the UI.

The `assignPostsToNewPromoter` cloud function has been refactored to be more robust and to correctly handle posts that should be assigned based on campaign name. It also ensures that only necessary fields are checked when querying for posts.